+++ packages/three-vrm-core/examples/models/cube.gltf
@@ -425,10 +425,10 @@
     }
   ],
   "extensionsUsed": [
-    "VRMC_vrm-1.0_draft"
+    "VRMC_vrm-1.0"
   ],
   "extensions": {
-    "VRMC_vrm-1.0_draft": {
+    "VRMC_vrm-1.0": {
       "meta": {
         "name": "Cube",
         "version": "1.0.0",
