+++ packages/three-vrm-core/src/firstPerson/index.ts
@@ -1,5 +1,4 @@
 export { VRMFirstPerson } from './VRMFirstPerson';
-export { VRMFirstPersonImporter } from './VRMFirstPersonImporter';
 export { VRMFirstPersonLoaderPlugin } from './VRMFirstPersonLoaderPlugin';
 export type { VRMFirstPersonMeshAnnotation } from './VRMFirstPersonMeshAnnotation';
 export { VRMFirstPersonMeshAnnotationType } from './VRMFirstPersonMeshAnnotationType';
