+++ packages/three-vrm-core/src/humanoid/index.ts
@@ -2,7 +2,6 @@ export { VRMHumanBone } from './VRMHumanBone';
 export { VRMHumanBoneName } from './VRMHumanBoneName';
 export type { VRMHumanBones } from './VRMHumanBones';
 export { VRMHumanoid } from './VRMHumanoid';
-export { VRMHumanoidImporter } from './VRMHumanoidImporter';
 export { VRMHumanoidLoaderPlugin } from './VRMHumanoidLoaderPlugin';
 export type { VRMPose } from './VRMPose';
 export type { VRMPoseTransform } from './VRMPoseTransform';
