+++ packages/three-vrm-core/src/lookAt/index.ts
@@ -2,7 +2,6 @@ export { VRMLookAt } from './VRMLookAt';
 export type { VRMLookAtApplier } from './VRMLookAtApplier';
 export { VRMLookAtBoneApplier } from './VRMLookAtBoneApplier';
 export { VRMLookAtExpressionApplier } from './VRMLookAtExpressionApplier';
-export { VRMLookAtImporter } from './VRMLookAtImporter';
 export { VRMLookAtLoaderPlugin } from './VRMLookAtLoaderPlugin';
 export { VRMLookAtRangeMap } from './VRMLookAtRangeMap';
 export { VRMLookAtTypeName } from './VRMLookAtTypeName';
