+++ packages/three-vrm-core/src/meta/VRMMetaLoaderPluginOptions.ts
@@ -1,7 +1,7 @@
 /**
- * A set of options for a {@link VRMMetaImporter} instance.
+ * A set of options for a {@link VRMMetaLoaderPlugin} instance.
  */
-export interface VRMMetaImporterOptions {
+export interface VRMMetaLoaderPluginOptions {
   /**
    * If `false`, it won't load its thumbnail image ({@link VRMMeta.thumbnailImage}). `true` by default.
    */
